+++ Reserch/Assets/Script/Command/putCube.cs
@@ -24,7 +24,7 @@ public class putCube : Command
         //Floor[,] map = GameObject.FindGameObjectWithTag("Map").GetComponent<Map>().getMap();
         Vector2Int putCubePos = CharacterScript.gameObject.getMapPosition() + new Vector2Int(1, 0);
         Vector3 characterPos = CharacterScript.gameObject.transform.position;
-        Vector3 putCubeV3Pos = characterPos + new Vector3(1.6f, 0, -1);
+        Vector3 putCubeV3Pos = characterPos + new Vector3(CO.FLOOR_DISTANCE.X, 0, -1);
 
 
         if (map[putCubePos.x, putCubePos.y].getGameObjectOnFloor() is null)
