+++ Reserch/Assets/Script/Object.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f9baa7f2cbbec6a42b6ee05f782f69f4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
