+++ Reserch/Assets/Script/Object/CharacterState.cs
