+++ Reserch/Assets/Script/Object/Cube.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Cube : Object
+{
+
+}
