+++ Reserch/Assets/Script/Object/Cube.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7be372304cf37e342afa7e6823056018
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
